Address SR04 hang problem.
Address SR04 hang problem where reading the GPIO Echo pin did not always give back a changed condition. Add a measure to also look for a change in the time.  

Created a SR04 Threaded Test module 

Finally, in the troubleshooting process, an SR04 version was created using the 'pigpio.py' library version. This did not address the original problem due to the nature of reading GPIOs for state and event changes.